feat: split out system() calls from the rest of MasterServer
